pipeline.sample returns 1 image if given 1 prompt
The example in the README for sample is:

  import torch
  from model import StableMultiDiffusionPipeline

  device = 0

  # Load the module.
  device = torch.device(f'cuda:{device}')
  smd = StableMultiDiffusionPipeline(device)

  # Sample an image.
  image = smd.sample('A photo of the dolomites')
  image.save('my_creation.png')

But this example does not work as-is because
currently `sample` always returns a list of images
since it might be given a list of prompts.

This change allows the example to work by supporting
this common case: a single image is now returnsed when
sample is given a single prompt.